requirements.txt
python-telegram-bot[job-queue]==22.6
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.28.1
h2==4.2.0
beautifulsoup4==4.12.3
pillow==12.1.0
telegraph==2.2.0
playwright==1.54.0
